renamed post-list-and-post .js componentsto better conform to my ideas.
