+++ src/scripts/data/DataManager.js
@@ -2,10 +2,27 @@ export const getPosts = () => {
     return fetch("http://localhost:8088/entries")
     .then(response => response.json())
 };
+
+
+
+
+//vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
+
+
 export const getPlaylistImages = () => {
     return fetch("http://localhost:8088/randomPlaylistArt")
     .then(response => response.json())
 };
+
+
+
+
+
+
+
+
+
+
 
 export const getSoundcloud = () => {
     return fetch("http://localhost:8088/playlist")
