+++ src/scripts/playlistImage/playlistImage.js
